+++ docs/semantic_annotation_parser_book-html/searcher-c2a407aa.js
@@ -437,7 +437,7 @@ window.search = window.search || {};
         if (yes) {
             loadSearchScript(
                 window.path_to_searchindex_js ||
-                path_to_root + 'searchindex-502eb58c.js',
+                path_to_root + 'searchindex-6cdfad2f.js',
                 'mdbook-search-index');
             search_wrap.classList.remove('hidden');
             searchicon.setAttribute('aria-expanded', 'true');